Added CRUD operations and category services
Added a new set of classes to handle CRUD operations as well as created services for category objects. The changes involve the addition of RestBase and CrudRestHelperBase classes that provide a general framework for performing CRUD operations. Additionally, added a new interface for CRUD operations and implemented specific operations for the category proxy. Also, a razor page to list categories is added.